pyproject.toml
[project]
name = "things-3-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "things-py>=0.0.15",
]
